+++ durastore/src/main/java/org/duracloud/durastore/rest/ContentResourceImpl.java
@@ -153,7 +153,7 @@ public class ContentResourceImpl implements ContentResource {
                              InputStream content,
                              String contentMimeType,
                              Map<String, String> userProperties,
-                             int contentSize,
+                             long contentSize,
                              String checksum,
                              String storeID)
     throws ResourceException, InvalidIdException {
